refactor dashboard modal widgets
